staking: rename 'MultiRewards' to 'MultiDistributor'
