+++ pkg/distributors/contracts/interfaces/IMultiDistributor.sol
@@ -16,7 +16,7 @@ pragma solidity ^0.7.0;
 
 import "@balancer-labs/v2-solidity-utils/contracts/openzeppelin/IERC20.sol";
 
-interface IMultiRewards {
+interface IMultiDistributor {
     function create(
         IERC20 stakingToken,
         IERC20 rewardsToken,
